move pages to p1 dir and fix tests
