+++ frontend/src/Home.js
@@ -4,18 +4,18 @@ import { Trans } from '@lingui/macro'
 import styled from '@emotion/styled'
 import { PhaseBanner } from './components/phase-banner'
 import { PageNotFound } from './PageNotFound'
-import { FileUploadPage } from './FileUploadPage'
-import { Suspectinfo } from './SuspectInfoYesNoPage'
-import { ScamInfoPage } from './ScamInfoPage'
-import { MoneyLostPage } from './MoneyLostPage'
-import { ConfirmationPage } from './ConfirmationPage'
-import { SuspectInfoPage } from './SuspectInfoPage'
-import { ContactInfoPage } from './ContactInfoPage'
-import { LoseMoney } from './MoneyLostYesNoPage'
-import { ThankYou } from './ThankYou'
-import { Willbecontacted } from './WillBeContactedYesNoPage'
-import { LandingPage } from './LandingPage'
-import { FilesToUpload } from './FileUploadYesNoPage'
+import { FileUploadPage } from './p1/FileUploadPage'
+import { Suspectinfo } from './p1/SuspectInfoYesNoPage'
+import { ScamInfoPage } from './p1/ScamInfoPage'
+import { MoneyLostPage } from './p1/MoneyLostPage'
+import { ConfirmationPage } from './p1/ConfirmationPage'
+import { SuspectInfoPage } from './p1/SuspectInfoPage'
+import { ContactInfoPage } from './p1/ContactInfoPage'
+import { LoseMoney } from './p1/MoneyLostYesNoPage'
+import { ThankYou } from './p1/ThankYou'
+import { Willbecontacted } from './p1/WillBeContactedYesNoPage'
+import { LandingPage } from './p1/LandingPage'
+import { FilesToUpload } from './p1/FileUploadYesNoPage'
 
 const Root = styled('div')`
   margin: 20pt;
@@ -31,18 +31,18 @@ export const Home = () => (
     <Screen>
       <Router>
         <PageNotFound default />
-        <LandingPage path="/" />
-        <ScamInfoPage path="/scaminfo" />
-        <Suspectinfo path="/suspectinfoquestion" />
-        <SuspectInfoPage path="/suspectinfo" />
-        <LoseMoney path="/moneylostquestion" />
-        <MoneyLostPage path="/moneylost" />
-        <FilesToUpload path="/uploadfilesquestion" />
-        <FileUploadPage path="/uploadfiles" />
-        <Willbecontacted path="/contactinfoquestion" />
-        <ContactInfoPage path="/contactinfo" />
-        <ConfirmationPage path="/confirmation" />
-        <ThankYou path="/thankyou" />
+        <LandingPage path="/p1" />
+        <ScamInfoPage path="/p1/scaminfo" />
+        <Suspectinfo path="/p1/suspectinfoquestion" />
+        <SuspectInfoPage path="/p1/suspectinfo" />
+        <LoseMoney path="/p1/moneylostquestion" />
+        <MoneyLostPage path="/p1/moneylost" />
+        <FilesToUpload path="/p1/uploadfilesquestion" />
+        <FileUploadPage path="/p1/uploadfiles" />
+        <Willbecontacted path="/p1/contactinfoquestion" />
+        <ContactInfoPage path="/p1/contactinfo" />
+        <ConfirmationPage path="/p1/confirmation" />
+        <ThankYou path="/p1/thankyou" />
       </Router>
     </Screen>
   </Root>
