--- frontend/src/MoneyLostYesNoPage.js
@@ -1,53 +1,0 @@
-/**@jsx jsx */
-import { jsx } from '@emotion/core'
-import { Trans } from '@lingui/macro'
-import { H1 } from './components/header'
-import { P } from './components/paragraph'
-import { ButtonLink } from './components/link'
-import { TrackPageViews } from './TrackPageViews'
-import { Container } from './components/container'
-import { Steps } from './components/stepper'
-import { Layout } from './components/layout'
-
-export const LoseMoney = () => (
-  <Layout>
-    <TrackPageViews />
-    <Container
-      display="flex"
-      width="90%"
-      flexDirection="row"
-      marginBottom="20px"
-    >
-      <Steps activeStep={1} steps={[{}, {}, {}, {}, {}]} />
-    </Container>
-    <H1>
-      <Trans>Tell us if money was lost</Trans>
-    </H1>
-    <P>
-      <Trans>Did you lose money as a result of the scam?</Trans>
-    </P>
-
-    <Container
-      display="flex"
-      maxWidth={['200px', null, '300px']}
-      flexDirection="row"
-      justifyContent="space-between"
-      marginBottom=""
-    >
-      <ButtonLink
-        width={['40px', null, '50px']}
-        textAlign="center"
-        to="/moneylost"
-      >
-        <Trans>Yes</Trans>
-      </ButtonLink>
-      <ButtonLink
-        width={['40px', null, '50px']}
-        textAlign="center"
-        to="/suspectinfoquestion"
-      >
-        <Trans>No</Trans>
-      </ButtonLink>
-    </Container>
-  </Layout>
-)
