--- frontend/src/__tests__/ConfirmationPage.test.js
@@ -1,35 +1,0 @@
-import React from 'react'
-import { render, cleanup } from '@testing-library/react'
-import { ApolloProvider } from 'react-apollo'
-import { ThemeProvider } from 'emotion-theming'
-import { I18nProvider } from '@lingui/react'
-import { ConfirmationPage } from '../ConfirmationPage'
-import theme from '../theme'
-import en from '../../locale/en/messages.js'
-
-const client = {
-  readQuery: () => ({
-    scamInfo: JSON.stringify({}),
-    lostMoney: JSON.stringify({}),
-    suspectInfo: JSON.stringify({}),
-    files: [],
-    contactInfo: JSON.stringify({}),
-  }),
-  writeData: jest.fn(),
-}
-
-describe('<ConfirmationPage />', () => {
-  afterEach(cleanup)
-
-  it('renders', () => {
-    render(
-      <ThemeProvider theme={theme}>
-        <ApolloProvider client={client}>
-          <I18nProvider language={'en'} catalogs={{ en }}>
-            <ConfirmationPage />
-          </I18nProvider>
-        </ApolloProvider>
-      </ThemeProvider>,
-    )
-  })
-})
