--- frontend/src/forms/__tests__/ContactInfoForm.test.js
@@ -1,53 +1,0 @@
-import React from 'react'
-import wait from 'waait'
-import { render, fireEvent, cleanup } from '@testing-library/react'
-import { MockedProvider } from 'react-apollo/test-utils'
-import { ApolloProvider } from 'react-apollo'
-import { ThemeProvider } from 'emotion-theming'
-import { I18nProvider } from '@lingui/react'
-import { ContactInfoForm } from '../ContactInfoForm'
-import en from '../../../locale/en/messages.js'
-import theme from '../../theme'
-
-const catalogs = { en }
-
-const client = {
-  readQuery: () => ({
-    contactInfo: JSON.stringify({}),
-  }),
-  writeData: jest.fn(),
-}
-
-const fillIn = (element, { with: value }) =>
-  fireEvent.change(element, { target: { value } })
-
-const clickOn = element => fireEvent.click(element)
-
-describe('<ContactInfoForm />', () => {
-  afterEach(cleanup)
-
-  it.only('calls the onSubmit function when the form is submitted', async () => {
-    const submitMock = jest.fn()
-
-    const { getByLabelText, getByText } = render(
-      <ThemeProvider theme={theme}>
-        <MockedProvider mocks={[]} addTypename={false}>
-          <I18nProvider language={'en'} catalogs={catalogs}>
-            <ApolloProvider client={client}>
-              <ContactInfoForm onSubmit={submitMock} />
-            </ApolloProvider>
-          </I18nProvider>
-        </MockedProvider>
-      </ThemeProvider>,
-    )
-
-    const inputNode = getByLabelText('Full name')
-    const nextButton = getByText(/Continue/i)
-
-    fillIn(inputNode, { with: 'Mallory' })
-    clickOn(nextButton)
-    await wait(0) // Wait for promises to resolve
-
-    expect(submitMock).toHaveBeenCalledTimes(1)
-  })
-})
