+++ frontend/src/p1/SuspectInfoYesNoPage.js
@@ -1,0 +1,55 @@
+/**@jsx jsx */
+import { jsx } from '@emotion/core'
+import { Trans } from '@lingui/macro'
+import { H1 } from '../components/header'
+import { P } from '../components/paragraph'
+import { ButtonLink } from '../components/link'
+import { TrackPageViews } from '../TrackPageViews'
+import { Container } from '../components/container'
+import { Steps } from '../components/stepper'
+import { Layout } from '../components/layout'
+
+export const Suspectinfo = () => (
+  <Layout>
+    <TrackPageViews />
+    <Container
+      display="flex"
+      width="90%"
+      flexDirection="row"
+      marginBottom="20px"
+    >
+      <Steps activeStep={2} steps={[{}, {}, {}, {}, {}]} />
+    </Container>
+    <H1>
+      <Trans>Add scammer details</Trans>
+    </H1>
+    <P>
+      <Trans>
+        Do you have any information about who might be responsible for the scam?
+      </Trans>
+    </P>
+
+    <Container
+      display="flex"
+      maxWidth={['200px', null, '300px']}
+      flexDirection="row"
+      justifyContent="space-between"
+      marginBottom=""
+    >
+      <ButtonLink
+        width={['40px', null, '50px']}
+        textAlign="center"
+        to="/suspectinfo"
+      >
+        <Trans>Yes</Trans>
+      </ButtonLink>
+      <ButtonLink
+        width={['40px', null, '50px']}
+        textAlign="center"
+        to="/uploadfilesquestion"
+      >
+        <Trans>No</Trans>
+      </ButtonLink>
+    </Container>
+  </Layout>
+)
